create room.js
