+++ backend/index.js
@@ -3,14 +3,15 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const http = require('http');
 const websocket = require('ws');
-const aws = require('aws-sdk');
+
 const morgan = require('morgan');
-const { authenticateUserMiddleware, validateJwtMiddleware, getTtl } = require('./lib/utils');
 const broadcastWs = require('./lib/broadcastWs');
+const { authenticateUserMiddleware, validateJwtMiddleware } = require('./lib/utils');
+const { getRoomList, createRoom, updateRoom } = require('./lib/room');
 
 
 // Server Config
-const docClient = new aws.DynamoDB.DocumentClient();
+
 const app = express();
 const server = http.createServer(app);
 const wss = new websocket.Server({
@@ -44,54 +45,14 @@ app.post(
 
 app.get(
   '/room',
-  (req, res, next) => {
-    docClient.scan({
-      TableName: 'Room',
-      ProjectionExpression: 'id, roomInfo',
-    }).promise()
-      .then((data) => {
-        if (data.Count > 0) {
-          res.send(data.Items);
-        } else {
-          res.send([]);
-        }
-      })
-      .catch(err => next(err));
-  },
+  getRoomList,
 );
 
 
 app.post(
   '/room/:roomId',
   validateJwtMiddleware,
-  (req, res, next) => {
-    const { roomId } = req.params;
-    docClient.get({
-      TableName: 'Room',
-      Key: {
-        id: roomId,
-      },
-    }).promise()
-      .then((data) => {
-        if (data.Item) {
-          throw new Error('');
-        }
-
-        const defaultItem = {
-          id: roomId,
-          bidSeq: '[]',
-          cacheTtl: getTtl(),
-          roomInfo: {},
-        };
-
-        return docClient.put({
-          TableName: 'Room',
-          Item: defaultItem,
-        }).promise();
-      })
-      .then(() => res.sendStatus(200))
-      .catch(err => next(err));
-  },
+  createRoom,
 );
 
 
@@ -99,25 +60,5 @@ app.put(
   '/room/:roomId',
   validateJwtMiddleware,
   bodyParser.json(),
-  (req, res, next) => {
-    const { roomId } = req.params;
-    const bidSeq = JSON.stringify(req.body.bidSeq);
-    docClient.update({
-      TableName: 'Room',
-      Key: {
-        id: roomId,
-      },
-      UpdateExpression: 'set bidSeq = :b, cacheTtl = :t',
-      ExpressionAttributeValues: {
-        ':b': bidSeq,
-        ':t': getTtl(),
-      },
-      ReturnValues: 'ALL_NEW',
-    }).promise()
-      .then(() => {
-        broadcastWs.broadcastRoom(wss, roomId, JSON.stringify({ bidSeq }));
-        res.sendStatus(200);
-      })
-      .catch(err => next(err));
-  },
+  updateRoom(wss),
 );
